# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/category": {
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of categories to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of categories to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in name and description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of categories with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all categories with optional filters",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new category (Admin only)",
        "tags": [
          "categories"
        ]
      }
    },
    "/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get category by ID",
        "tags": [
          "categories"
        ]
      },
      "put": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Category not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a category (Admin only)",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Category not found"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a category (Admin only)",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/comment": {
      "get": {
        "operationId": "CommentController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of comments to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of comments to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "postId",
            "required": false,
            "in": "query",
            "description": "Filter by post ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentCommentId",
            "required": false,
            "in": "query",
            "description": "Filter by parent comment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of comments with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all comments with optional filters",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Cannot comment on a locked post"
          },
          "404": {
            "description": "Post not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new comment",
        "tags": [
          "comments"
        ]
      }
    },
    "/comment/{id}": {
      "get": {
        "operationId": "CommentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return comment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get comment by ID",
        "tags": [
          "comments"
        ]
      },
      "put": {
        "operationId": "CommentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to update comment"
          },
          "404": {
            "description": "Comment not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a comment",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to delete comment"
          },
          "404": {
            "description": "Comment not found"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string",
            "description": "User role"
          }
        },
        "required": [
          "_id",
          "email",
          "username",
          "password",
          "fullname",
          "role"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the category",
            "example": "Technology"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the category",
            "example": "Technology"
          }
        }
      },
      "Vote": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của lượt bình chọn"
          },
          "user": {
            "description": "Người dùng thực hiện bình chọn",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của người dùng thực hiện bình chọn"
          },
          "targetType": {
            "enum": [
              "post",
              "comment"
            ],
            "type": "string",
            "description": "Loại đối tượng được bình chọn (bài viết hoặc bình luận)",
            "example": "post"
          },
          "targetId": {
            "type": "string",
            "description": "ID của đối tượng được bình chọn"
          },
          "voteValue": {
            "enum": [
              1,
              -1
            ],
            "type": "number",
            "description": "Giá trị bình chọn (1 cho upvote, -1 cho downvote)",
            "example": 1
          },
          "post": {
            "description": "Bài viết được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "comment": {
            "description": "Bình luận được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình chọn"
          }
        },
        "required": [
          "_id",
          "user",
          "userId",
          "targetType",
          "targetId",
          "voteValue",
          "createdAt"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bình luận"
          },
          "post": {
            "description": "Bài viết chứa bình luận này",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "postId": {
            "type": "string",
            "description": "ID của bài viết"
          },
          "user": {
            "description": "Tác giả bình luận",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          },
          "parentComment": {
            "description": "Bình luận cha (nếu là bình luận trả lời)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "parentCommentId": {
            "type": "string",
            "description": "ID của bình luận cha (nếu có)"
          },
          "replies": {
            "description": "Các bình luận trả lời",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bình luận",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình luận"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bình luận"
          }
        },
        "required": [
          "_id",
          "post",
          "postId",
          "user",
          "userId",
          "content",
          "replies",
          "votes",
          "createdAt",
          "updatedAt"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bài viết"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "user": {
            "description": "Tác giả bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "category": {
            "description": "Danh mục của bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID của danh mục"
          },
          "tags": {
            "description": "Các thẻ được gắn cho bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "comments": {
            "description": "Các bình luận của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "isLocked": {
            "type": "boolean",
            "description": "Trạng thái khóa của bài viết"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bài viết"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bài viết"
          }
        },
        "required": [
          "_id",
          "title",
          "content",
          "user",
          "userId",
          "category",
          "categoryId",
          "tags",
          "comments",
          "votes",
          "isLocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của thẻ"
          },
          "tagName": {
            "type": "string",
            "description": "Tên của thẻ",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "Mô tả của thẻ"
          },
          "posts": {
            "description": "Các bài viết được gắn thẻ này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          }
        },
        "required": [
          "_id",
          "tagName",
          "posts"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "description": "ID của bài viết"
          },
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          },
          "parentCommentId": {
            "type": "string",
            "description": "ID của bình luận cha (nếu là bình luận trả lời)"
          }
        }
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          }
        }
      },
...
    }
  }
}